feat: amethyst golem tweaks
makes the golems not respect claims - current behaviour causes them to
drop the items but not break the block which results in an unintentional
buff.

also changes their harvesting behaviour:

if they do not have a home (not sure if even possible) or their home is
not an inventory, they will drop the drops as item entities.

if they do have a home and it is an inventory, they will insert the
dropped items directly into the home if there is free space or skip
harvesting the cluster otherwise. it will continue trying other clusters
in the latter case.